+++ R/raadtools.R
@@ -11,7 +11,6 @@
 #' @import methods
 #' @importFrom raster overlay
 #' @importFrom sp as.image.SpatialGridDataFrame bbox CRS GridTopology  proj4string<- SpatialPoints SpatialPointsDataFrame spChFIDs spTransform 
-#' @importFrom maptools ContourLines2SLDF spRbind
 #' @importFrom raster brick crop deratify extent<- extract getZ nlayers projection projection<- raster res resample rotate setZ stack writeRaster xmax xmin ymax ymin
 #' @details
 #'    read functions like \code{\link{readsst}} will read a data set by date-time vector, with a