+++ R/zzz.R
@@ -11,8 +11,6 @@
   
   options(raadtools.geoid_tile_vrt = NULL)
   
-  options(raadtools.do_terra = FALSE)
-  options(raadtools.be_terra = FALSE)
   ## send this message once per session
   options("raadtools.both.hemisphere.message" = FALSE)
 }